requirements.txt
pygame-ce==2.5.7
numpy==2.2.6
matplotlib==3.10.8
